CODAP-740 move formula.ts and formula-observers.ts
