+++ formulas/src/models/data/data-set-types.ts
@@ -1,8 +1,17 @@
-export interface IItem {
+import { IValueType } from "./attribute-types";
+
+export interface IItemID {
   __id__: string;
 }
 
-export interface ICase extends IItem {}
+export interface ICaseID extends IItemID {
+}
+
+export interface IItem extends IItemID {
+  [key: string]: IValueType;
+}
+export interface ICase extends IItem {
+}
 
 // used in IGroupedCase
 export const symParent = Symbol.for("parent")
