+++ formulas/todo.md
@@ -27,6 +27,7 @@ Files "symlinked":
 - functions/function-utils
 - math.ts
 - utils/name-mapping-utils.ts
+- formula.ts
 
 # Generics
 
@@ -63,6 +64,7 @@ These cast the host global value manager to the formula global value manager:
   - formula-test-utils.ts
 
 # Things to fix:
+- formula-types.ts still exists in v3
 - the scss files should not be processed, just copied to dist, so then when the v3 package bundles the dependencies it will process these scss files and update the imports appropriately.
 
 # Type issues
